refactor: extract issue_cache and tool_defs from mcp_channel.ts
Reduces mcp_channel.ts from 565 to 370 lines.
Adds role, workdir, project metadata to agent registration.
New env: SCITEX_OROCHI_ROLE, SCITEX_OROCHI_PROJECT.
